+++ gui/MainWindowCentralWidget.py
@@ -346,6 +346,7 @@ class MainWindowCentralWidget(QWidget):
                                                     IDESettings.get_font_boldness())
                 tab.supply_editor_color_theme_to_lexer()
                 tab.setup_autocompletion()
+                tab.set_highlight_style(IDESettings.get_highlight_fill_rect())
 
     def handle_insert_into_cursor(self, command: str):
         # Don't do anything if there are no tabs open